12x 1-column bug - fixed!
The bug has been fixed simply by removing the 1px border, which was
there only to visualize the columns. Added a background color instead.

The pixel size has been converted and replaced with 'em' based values
instead. Due to this, the calculations within the algorithm has been
optimized, giving a better result in the browser.

TODO: Mobile-first design, and @media-queries in general.